Remove Mogenerator

+++ ChatSecure/OTREncryptionManager.m
@@ -215,7 +215,7 @@ NSString *const OTRMessageStateKey = @"OTREncryptionManagerMessageStateKey";
 
 - (void) otrKit:(OTRKit *)otrKit handleMessageEvent:(OTRKitMessageEvent)event message:(NSString *)message username:(NSString *)username accountName:(NSString *)accountName protocol:(NSString *)protocol tag:(id)tag error:(NSError *)error {
     //incoming and outgoing errors and other events
-    DDLogWarn(@"Message Event: %lu Error:%@",event,error);
+    DDLogWarn(@"Message Event: %d Error:%@",(int)event,error);
     
     if ([tag isKindOfClass:[OTRMessage class]]) {
         __block NSError *error = nil;
